federated-graph: remove VersionedFederatedGraph
There is no longer a need for versioning with the removal of managed graphs. Use the concrete FederatedGraph type everywhere, and rename the federated_graph::v4 module to federated_graph.

closes GB-7948
