+++ engine/crates/composition/tests/composition_tests.rs
@@ -46,13 +46,10 @@ fn run_test(federated_graph_path: &Path) -> datatest_stable::Result<()> {
         .map_err(|err| miette::miette!("Error trying to read api.graphql: {}", err))
         .ok();
     let (actual_federated_sdl, actual_api_sdl) = match graphql_composition::compose(&subgraphs).into_result() {
-        Ok(federated_graph) => {
-            let federated_graph = federated_graph.into_latest().expect("federated graph into_latest()");
-            (
-                graphql_federated_graph::render_federated_sdl(&federated_graph).expect("rendering federated SDL"),
-                Some(graphql_federated_graph::render_api_sdl(&federated_graph)),
-            )
-        }
+        Ok(federated_graph) => (
+            graphql_federated_graph::render_federated_sdl(&federated_graph).expect("rendering federated SDL"),
+            Some(graphql_federated_graph::render_api_sdl(&federated_graph)),
+        ),
         Err(diagnostics) => (
             format!(
                 "{}\n",
