+++ engine/crates/federated-graph/src/render_sdl/render_federated_sdl.rs
@@ -636,11 +636,9 @@ mod tests {
     fn test_render_empty() {
         use expect_test::expect;
 
-        let empty = crate::VersionedFederatedGraph::Sdl(
-            crate::render_sdl::render_federated_sdl(&FederatedGraph::default()).unwrap(),
-        );
+        let empty = FederatedGraph::default();
 
-        let actual = render_federated_sdl(&empty.into_latest().unwrap()).expect("valid");
+        let actual = render_federated_sdl(&empty).expect("valid");
         let expected = expect![[r#"
             directive @core(feature: String!) repeatable on SCHEMA
