+++ gateway/crates/gateway-binary/tests/telemetry/tracing.rs
@@ -391,11 +391,8 @@ where
         subgraphs
             .ingest_str(&subgraph_sdl, "the-subgraph", subgraph_server.url().as_str())
             .unwrap();
-        graphql_composition::compose(&subgraphs)
-            .into_result()
-            .unwrap()
-            .into_federated_sdl()
-            .expect("graph.into_latest()")
+        let graph = graphql_composition::compose(&subgraphs).into_result().unwrap();
+        graphql_composition::render_federated_sdl(&graph).unwrap()
     };
 
     crate::GatewayBuilder {
